trained and tested A - E asl
I've trained the model for multiple hand gesture A - E in the auslan fingerspelling. There is one letter C which may collide with asl but we're here to find out. This repo is very close to being used to train both asl and auslan, just little changes need to be made.

https://auslan.org.au/spell/twohanded.html